Cool and funny solution for IE9 support.
IE9 doesn't support event propagation through the canvas instance therefore it is necessary to re-send event to the each element:
http://stackoverflow.com/questions/7446720/ie9-forward-mouse-events-through-a-masking-canvas-element

But I have implemented and alternative solution:
1. Check if there is any element available under canvas on mouse move $.("#" + euid +"_Canvas").mousemove(function(event) { diagram.isMouseOverElement(event.point);})
2. Select element with the biggest z-index at this point and "+=3000" z-index for it
3. Make element semi-transparent on mouse enter
4. And make it not-transparent on mouse leave;

So, it seems that it is bloody solution work :) Of course with minor issues which required further analysis. But I think that "prove of concept" is done.